requirements.txt
# API & Authentication
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Data Processing
pandas>=2.0.0

# Google Cloud
google-cloud-bigquery>=3.11.0

# Development & Testing
python-dotenv>=1.0.0
pytest>=7.4.0